requirements.txt
torch
torchvision
transformers
diffusers
accelerate
pillow
numpy
gradio==3.41.2
huggingface-hub
